requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2>=2.8
djangorestframework-api-key==2.*
djangorestframework-simplejwt
requests